requirements.txt
openai==1.13.3
chromadb==0.4.24
numpy==1.26.4
httpx==0.27.0
gitpython==3.1.42
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
tiktoken==0.6.0